+++ AndriodApp1/Chatroom/ChatroomActivity.cs
@@ -19,6 +19,7 @@
 
 using AndriodApp1.Chatroom;
 using AndriodApp1.Helpers;
+using AndriodApp1.Messages;
 using Android.App;
 using Android.Content;
 using Android.Graphics;
