+++ AndriodApp1/Chatroom/ChatroomController.cs
@@ -4,10 +4,12 @@ using Android.OS;
 using Android.Runtime;
 using Android.Views;
 using Android.Widget;
+using AndroidX.Core.App;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Threading.Tasks;
 
 namespace AndriodApp1.Chatroom
 {
