+++ AndriodApp1/Chatroom/RoomUserListDialog.cs
@@ -1,6 +1,8 @@
 ﻿using AndriodApp1.Helpers;
+using AndriodApp1.Messages;
 using Android.App;
 using Android.Content;
+using Android.Graphics;
 using Android.OS;
 using Android.Runtime;
 using Android.Views;
