+++ AndriodApp1/Serialization/SerializationHelper.cs
@@ -1,4 +1,5 @@
-﻿using Android.Hardware.Camera2;
+﻿using AndriodApp1.Messages;
+using Android.Hardware.Camera2;
 using AndroidX.DocumentFile.Provider;
 using System;
 using System.Collections.Concurrent;
