+++ AndriodApp1/Users/ViewUserInfoActivity.cs
@@ -18,6 +18,7 @@
  */
 
 using AndriodApp1.Helpers;
+using AndriodApp1.Messages;
 using Android.App;
 using Android.Content;
 using Android.Graphics;
